+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## v0.1.18
 * Added: `NativeStylesController.applyStyles(.., priorityLevel)` Higher style levels will override styles defined at lower levels. Default level is 0.
-* Added: `NativeStylesController.getAppliedStyles`
+* Added: `NativeStylesController.getAppliedStyles()`
+* Added: `UIComponent.getAppliedStyles()`
 
 ## v0.1.17
 * Added: `ColorUtils.addAlphaToCSS`